+++ src/index.ts
@@ -3,6 +3,7 @@ export * from "./core/DbContext";
 export * from "./core/DbSet";
 export * from "./core/MetadataStorage";
 export * from "./core/ModelBuilder";
+export * from "./core/QueryFilter";
 export * from "./core/EntityEntry";
 export * from "./core/ChangeTracker";
 export * from "./core/utils";